scripts/pyproject.toml
[tool.poetry]
name = "lyahelia-art"
version = "0.1.0"
description = "Lya Helia Art Website"
authors = ["Lya Helia"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^11.1.0"
pillow-heif = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
